d_camera work (#2369)
* bumpCheck matching

* checkGroundInfo matching

* chaseCamera matching

* lockonCamera attempted

* talktoCamera matching

* subjectCamera, magneCamera matching

* colosseumCamera close, plus various fixes

* towerCamera attempted

* hookshotCamera matching

+++ include/d/d_stage.h
@@ -561,7 +561,7 @@ public:
 
     void setTreasure(stage_tresure_class*);
 
-    s16 getWorldRollAngleX() { return mWorldRollAngleX; }
+    s16 getWorldRollAngleX() { return (s16)mWorldRollAngleX; }
     s16 getWorldRollDirAngleY() { return mWorldRollDirAngleY; }
 
 public:
